+++ src/types/index.ts
@@ -1,2 +1,4 @@
+export * from './defaultToken.types'
+export * from './icons.types'
 export * from './internal.types'
 export * from '@lifinance/types'
